+++ ChangeLog.md
@@ -41,6 +41,9 @@ the value of this new parameter.
 7. Fixed a regression introduced by 3.0 beta1[24] that caused the TurboVNC
 Server to leak memory when the RFB flow control extensions were used.
 
+8. Fixed several issues in the TurboVNC Viewer that prevented IPv6 addresses
+from being used as VNC hostnames.
+
 
 3.1.1
 =====